pyproject.toml
[project]
name = "noteme"
version = "0.1.0"
description = "Telegram bot for tracking important dates, beautiful milestones, and notes"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Telegram
    "aiogram>=3.20",
    # Web
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    # Database
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    # Settings
    "pydantic-settings>=2.7",
    # Cache / Queue
    "redis>=5.2",
    "arq>=0.26",
    # AI
    "langgraph>=0.4",
    "langchain-openai>=0.3",
    "openai>=1.60",
    # Admin
    "sqladmin>=0.20",
    # Date calculations
    "python-dateutil>=2.9",
    # Declensions
    "pymorphy3>=2.0",
    "inflect>=7.4",
    # Templates
    "jinja2>=3.1",
    # Monitoring
    "sentry-sdk[fastapi]>=2.19",
    "prometheus-client>=0.21",
    # HTTP client
    "httpx>=0.28",
    "itsdangerous>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.25",
    "pytest-cov>=6.0",
    "ruff>=0.9",
    "mypy>=1.14",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]
